Continuous and discrete steps, assuming only top level parallelism.
Arbitrary parallelism is not theoretically any worse, but it causes
more complexity explosion in the flattened model, I'm not sure it adds
much useful expressiveness, and it makes the semantics of transitions
much more confusing.  The schema supports arbitrary parallelism but
this reference interpreter only accepts models with top-level
parallelism.  Most of the interpreter is agnostic, just the code for
discrete transitions and its supporting code has to deal with it.