Propagate some entry conditions into the invariant.
Specifically, we take the disjunction of incoming edge conditions,
where an edge condition is:

* The updates to variables on that edge
* Guard terms (elements in clauses in the DNF representation of the
  guard) which happen not to use lag variables or time.
** E.g. y'<=0 lets us say y'_<=0 in the new state
* Guard terms involving collisions let us put v'_ == 0 constraints on
  the corresponding normals (or disjunction of both if no normals are
  given).
* Other kinds of propagation are possible, e.g. considering the flows
  or invariants of the entering states.

NOTE: this will not look at edges into a parent state which
default-enters this state.  That's a potential bug, since we assume we
get to see all the enter conditions.

In the end, I'm going to put a flag on this because it's cool but
during planning it's not really necessary if we can use the solver
during planning; it remains to be seen if that's a bad idea.