Fix problems with entry guard propagation into invariant.
This seems to be accurate now; also it lets us compute that e.g. mario
jump_fixed will never be active for less than 0.7 seconds given the
default parameters!  So that's cool.  I think with a bit more work we
could learn an upper bound on that time too.  This all should
generalize to other modes maybe?